+++ galsim/hsm.py
@@ -76,18 +76,20 @@ class ShapeData(object):
     - observed_shape: a Shear object representing the observed shape based on adaptive
       moments.
 
-    - moments_sigma: size sigma = (det M)^(1/4) from the adaptive moments, in units of pixels; -1 if
+    - moments_sigma: size sigma=(det M)^(1/4) from the adaptive moments, in units of pixels; -1 if
       not measured.
 
     - moments_amp: total image intensity for best-fit elliptical Gaussian from adaptive moments.
       Normally, this field is simply equal to the image flux (for objects that follow a Gaussian
       light distribution, otherwise it is something approximating the flux).  However, if the image
       was drawn using `drawImage(method='sb')` then moments_amp relates to the flux via
-      flux = (moments_amp)*(pixel scale)^2.
+      flux=(moments_amp)*(pixel scale)^2.
 
-    - moments_centroid: a PositionD object representing the centroid based on adaptive
-      moments.  The convention for centroids is such that the center of the lower-left pixel is
-      (0,0).
+    - moments_centroid: a PositionD object representing the centroid based on adaptive moments, in
+      units of pixels.  The indexing convention is defined with respect to the BoundsI object
+      defining the bounds of the input Image, i.e., the center of the lower left pixel is
+      `(image.xmin, image.ymin)`.  An object drawn at the center of the image should generally have
+      moments_centroid equal to image.trueCenter().
 
     - moments_rho4: the weighted radial fourth moment of the image.
 
@@ -112,6 +114,11 @@ class ShapeData(object):
 
     - resolution_factor: Resolution factor R_2;  0 indicates object is consistent with a PSF, 1
       indicates perfect resolution.
+
+    - psf_sigma: size sigma=(det M)^(1/4) of the PSF from the adaptive moments, in units of pixels;
+      -1 if not measured.
+
+    - psf_shape: a Shear object representing the observed PSF shape based on adaptive moments.
 
     - error_message: a string containing any error messages from the attempt to carry out
       PSF-correction.
@@ -144,6 +151,8 @@ class ShapeData(object):
             self.corrected_shape_err = args[0].corrected_shape_err
             self.correction_method = args[0].correction_method
             self.resolution_factor = args[0].resolution_factor
+            self.psf_sigma = args[0].psf_sigma
+            self.psf_shape = galsim.Shear(args[0].psf_shape)
             self.error_message = args[0].error_message
         else:
             self.image_bounds = _galsim.BoundsI()
@@ -163,11 +172,13 @@ class ShapeData(object):
             self.corrected_shape_err = -1.0
             self.correction_method = "None"
             self.resolution_factor = -1.0
+            self.psf_sigma = -1.0
+            self.psf_shape = galsim.Shear()
             self.error_message = ""
 
 # A helper function for taking input weight and badpix Images, and returning a weight Image in the
 # format that the C++ functions want
-def _convertMask(image, weight = None, badpix = None):
+def _convertMask(image, weight=None, badpix=None):
     """Convert from input weight and badpix images to a single mask image needed by C++ functions.
 
     This is used by EstimateShear() and FindAdaptiveMom().
@@ -215,10 +226,10 @@ def _convertMask(image, weight = None, badpix = None):
     # finally, return the Image for the weight map
     return mask.image.view()
 
-def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 0.0,
-                  shear_est = "REGAUSS", recompute_flux = "FIT", guess_sig_gal = 5.0,
-                  guess_sig_PSF = 3.0, precision = 1.0e-6, guess_x_centroid = -1000.0,
-                  guess_y_centroid = -1000.0, strict = True, hsmparams = None):
+def EstimateShear(gal_image, PSF_image, weight=None, badpix=None, sky_var=0.0,
+                  shear_est="REGAUSS", recompute_flux="FIT", guess_sig_gal=5.0,
+                  guess_sig_PSF=3.0, precision=1.0e-6, guess_centroid=None,
+                  strict=True, hsmparams=None):
     """Carry out moments-based PSF correction routines.
 
     Carry out PSF correction using one of the methods of the HSM package (see references in
@@ -230,6 +241,9 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
     corresponds to no particular GSObject), but this also means that users who wish to apply it to
     simple combinations of GSObjects (e.g., a Convolve) must take the additional step of drawing
     their GSObjects into Images.
+
+    This routine assumes that (at least locally) the WCS can be approximated as a PixelScale, with
+    no distortion or non-trivial remapping. Any non-trivial WCS gets completely ignored.
 
     Note that the method will fail if the PSF or galaxy are too point-like to easily fit an
     elliptical Gaussian; when running on batches of many galaxies, it may be preferable to set
@@ -243,19 +257,18 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
 
     Typical application to a single object:
 
-        >>> galaxy = galsim.Gaussian(flux = 1.0, sigma = 1.0)
+        >>> galaxy = galsim.Gaussian(flux=1.0, sigma=1.0)
         >>> galaxy = galaxy.shear(g1=0.05, g2=0.0)  # shears the Gaussian by (0.05, 0) using the 
-        >>>                                         # |g| = (a - b)/(a + b) definition
-        >>> psf = galsim.Kolmogorov(flux = 1.0, fwhm = 0.7)
-        >>> final = galsim.Convolve([galaxy, psf])
-        >>> final_image = final.drawImage(dx = 0.2)
-        >>> final_epsf_image = psf.drawImage(dx = 0.2)
+        >>>                                         # |g| = (a-b)/(a+b) definition
+        >>> psf = galsim.Kolmogorov(flux=1.0, fwhm=0.7)
+        >>> final = galsim.Convolve(galaxy, psf)
+        >>> final_image = final.drawImage(scale=0.2)
+        >>> final_epsf_image = psf.drawImage(scale=0.2)
         >>> result = galsim.hsm.EstimateShear(final_image, final_epsf_image)
     
-    After running the above code, `result.observed_shape` ["shape" = distortion, the 
-    (a^2 - b^2)/(a^2 + b^2) definition of ellipticity] is
-    `(0.0438925351523, -1.12519277137e-18)` and `result.corrected_e1`, `result_corrected_e2` are
-    `(0.0993412658572,-1.84832327221e-09)`, compared with the  expected `(0.09975, 0)` for a perfect
+    After running the above code, `result.observed_shape` is a galsim.Shear object with a value of
+    `(0.0438925349133, -2.85747392701e-18)` and `result.corrected_e1`, `result_corrected_e2` are
+    `(0.09934103488922119, -3.746108423463568e-10)`, compared with the expected `(0.09975, 0)` for a perfect
     PSF correction method.
 
     The code below gives an example of how one could run this routine on a large batch of galaxies,
@@ -265,7 +278,7 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
         >>> n_fail = 0
         >>> for i=0, range(n_image):
         >>>     #...some code defining this_image, this_final_epsf_image...
-        >>>     result = galsim.hsm.EstimateShear(this_image, this_final_epsf_image, strict = False)
+        >>>     result = galsim.hsm.EstimateShear(this_image, this_final_epsf_image, strict=False)
         >>>     if result.error_message != "":
         >>>         n_fail += 1
         >>> print "Number of failures: ", n_fail
@@ -294,15 +307,12 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
     @param guess_sig_PSF    Optional argument with an initial guess for the Gaussian sigma of the
                             PSF (in pixels). [default: 3.]
     @param precision        The convergence criterion for the moments. [default: 1e-6]
-    @param guess_x_centroid  An initial guess for the x component of the object centroid (useful in
-                            case it is not located at the center, which is the default
-                            assumption).  The convention for centroids is such that the center of
-                            the lower-left pixel is (0,0). [default: gal_image.trueCenter().x]
-    @param guess_y_centroid  An initial guess for the y component of the object centroid (useful in
-                            case it is not located at the center, which is the default
-                            assumption).  The convention for centroids is such that the center of
-                            the lower-left pixel is (0,0). [default: gal_image.trueCenter().y]
-    @param strict           Whether to require success. If `strict = True`, then there will be a 
+    @param guess_centroid   An initial guess for the object centroid (useful in
+                            case it is not located at the center, which is used if this keyword is
+                            not set).  The convention for centroids is such that the center of
+                            the lower-left pixel is (image.xmin, image.ymin).
+                            [default: gal_image.trueCenter()]
+    @param strict           Whether to require success. If `strict=True`, then there will be a 
                             `RuntimeError` exception if shear estimation fails.  If set to `False`,
                             then information about failures will be silently stored in the output 
                             ShapeData object. [default: True]
@@ -317,6 +327,8 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
     PSF_image_view = PSF_image.image.view()
     weight_view = _convertMask(gal_image, weight=weight, badpix=badpix)
 
+    if guess_centroid is None:
+        guess_centroid = gal_image.trueCenter()
     try:
         result = _galsim._EstimateShearView(gal_image_view, PSF_image_view, weight_view,
                                             sky_var = sky_var,
@@ -325,8 +337,7 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
                                             guess_sig_gal = guess_sig_gal,
                                             guess_sig_PSF = guess_sig_PSF,
                                             precision = precision,
-                                            guess_x_centroid = guess_x_centroid,
-                                            guess_y_centroid = guess_y_centroid,
+                                            guess_centroid = guess_centroid,
                                             hsmparams = hsmparams)
     except RuntimeError as err:
         if (strict == True):
@@ -336,9 +347,8 @@ def EstimateShear(gal_image, PSF_image, weight = None, badpix = None, sky_var = 
             result.error_message = str(err)
     return ShapeData(result)
 
-def FindAdaptiveMom(object_image, weight = None, badpix = None, guess_sig = 5.0, precision = 1.0e-6,
-                    guess_x_centroid = -1000.0, guess_y_centroid = -1000.0, strict = True,
-                    hsmparams = None):
+def FindAdaptiveMom(object_image, weight=None, badpix=None, guess_sig=5.0, precision=1.0e-6,
+                    guess_centroid=None, strict=True, hsmparams=None):
     """Measure adaptive moments of an object.
 
     This method estimates the best-fit elliptical Gaussian to the object (see Hirata & Seljak 2003
@@ -349,14 +359,17 @@ def FindAdaptiveMom(object_image, weight = None, badpix = None, guess_sig = 5.0,
     weight function.  FindAdaptiveMom() can be used either as a free function, or as a method of the
     Image class.
 
+    This routine assumes that (at least locally) the WCS can be approximated as a PixelScale, with
+    no distortion or non-trivial remapping. Any non-trivial WCS gets completely ignored.
+
     Like EstimateShear(), FindAdaptiveMom() works on Image inputs, and fails if the object is small
     compared to the pixel scale.  For more details, see EstimateShear().
 
     Example usage
     -------------
 
-        >>> my_gaussian = galsim.Gaussian(flux = 1.0, sigma = 1.0)
-        >>> my_gaussian_image = my_gaussian.drawImage(dx = 0.2, method='no_pixel')
+        >>> my_gaussian = galsim.Gaussian(flux=1.0, sigma=1.0)
+        >>> my_gaussian_image = my_gaussian.drawImage(scale=0.2, method='no_pixel')
         >>> my_moments = galsim.hsm.FindAdaptiveMom(my_gaussian_image)
 
     OR
@@ -365,21 +378,20 @@ def FindAdaptiveMom(object_image, weight = None, badpix = None, guess_sig = 5.0,
 
     Assuming a successful measurement, the most relevant pieces of information are
     `my_moments.moments_sigma`, which is `|det(M)|^(1/4)` [=`sigma` for a circular Gaussian] and
-    `my_moments.observed_shape`, which is a Shear.  
+    `my_moments.observed_shape`, which is a Shear.  In this case, `my_moments.moments_sigma` is
+    precisely 5.0 (in units of pixels), and `my_moments.observed_shape` is consistent with zero.
 
-    Methods of the Shear class can be used to get the distortion
-    `(e1, e2) = (a^2 - b^2)/(a^2 + b^2)`, e.g. `my_moments.observed_shape.getE1()`, or to get the
-    shear `g`, the conformal shear `eta`, and so on.
+    Methods of the Shear class can be used to get the distortion `e`, the shear `g`, the conformal
+    shear `eta`, and so on.
 
     As an example of how to use the optional `hsmparams` argument, consider cases where the input
     images have unusual properties, such as being very large.  This could occur when measuring the
     properties of a very over-sampled image such as that generated using
 
-        >>> my_gaussian = galsim.Gaussian(sigma = 5.0)
-        >>> my_gaussian_image = my_gaussian.drawImage(dx = 0.01, method='no_pixel')
+        >>> my_gaussian = galsim.Gaussian(sigma=5.0)
+        >>> my_gaussian_image = my_gaussian.drawImage(scale=0.01, method='no_pixel')
 
-    If the user attempts to measure the moments of this 4000 x 4000 pixel image using the standard
-    syntax,
+    If the user attempts to measure the moments of this very large image using the standard syntax,
 
         >>> my_moments = my_gaussian_image.FindAdaptiveMom()
 
@@ -391,7 +403,7 @@ def FindAdaptiveMom(object_image, weight = None, badpix = None, guess_sig = 5.0,
     successful:
 
         >>> new_params = galsim.hsm.HSMParams(max_amoment=5.0e5)
-        >>> my_moments = my_gaussian_image.FindAdaptiveMom(hsmparams = new_params)
+        >>> my_moments = my_gaussian_image.FindAdaptiveMom(hsmparams=new_params)
 
     @param object_image     The Image for the object being measured.
     @param weight           The optional weight image for the object being measured.  Can be an int
@@ -406,15 +418,12 @@ def FindAdaptiveMom(object_image, weight = None, badpix = None, guess_sig = 5.0,
     @param guess_sig        Optional argument with an initial guess for the Gaussian sigma of the
                             object (in pixels). [default: 5.0]
     @param precision        The convergence criterion for the moments. [default: 1e-6]
-    @param guess_x_centroid  An initial guess for the x component of the object centroid (useful in
-                            case it is not located at the center, which is the default
-                            assumption).  The convention for centroids is such that the center of
-                            the lower-left pixel is (0,0). [default: gal_image.trueCenter().x]
-    @param guess_y_centroid  An initial guess for the y component of the object centroid (useful in
-                            case it is not located at the center, which is the default
-                            assumption).  The convention for centroids is such that the center of
-                            the lower-left pixel is (0,0). [default: gal_image.trueCenter().y]
-    @param strict           Whether to require success. If `strict = True`, then there will be a 
+    @param guess_centroid   An initial guess for the object centroid (useful in case it is not
+                            located at the center, which is used if this keyword is not set).  The
+                            convention for centroids is such that the center of the lower-left pixel
+                            is (image.xmin, image.ymin).
+                            [default: object_image.trueCenter()]
+    @param strict           Whether to require success. If `strict=True`, then there will be a 
                             `RuntimeError` exception if shear estimation fails.  If set to `False`,
                             then information about failures will be silently stored in the output 
                             ShapeData object. [default: True]
@@ -428,11 +437,13 @@ def FindAdaptiveMom(object_image, weight = None, badpix = None, guess_sig = 5.0,
     object_image_view = object_image.image.view()
     weight_view = _convertMask(object_image, weight=weight, badpix=badpix)
 
+    if guess_centroid is None:
+        guess_centroid = object_image.trueCenter()
+
     try:
         result = _galsim._FindAdaptiveMomView(object_image_view, weight_view,
                                               guess_sig = guess_sig, precision =  precision,
-                                              guess_x_centroid = guess_x_centroid,
-                                              guess_y_centroid = guess_y_centroid,
+                                              guess_centroid = guess_centroid,
                                               hsmparams = hsmparams)
     except RuntimeError as err:
         if (strict == True):